backend/requirements.txt
#app

fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
redis==5.2.0
httpx==0.27.2
authlib==1.3.2
python-multipart==0.0.19
orjson==3.10.12
celery==5.4.0






#preview_workers:

celery==5.4.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.1
redis==5.2.0
python-jose[cryptography]==3.3.0



ai_workers:

celery==5.4.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.1
redis==5.2.0
